+++ Assets/3 Scripts/4 Game/Enemy/Enemy_Honeycomb.cs
@@ -52,7 +52,7 @@ public class Enemy_Honeycomb : NonDamagableEnemy
         shootTimer = 0;
     }
 
-    public override int TakeDamage(int attckAmount)
+    public override float TakeDamage(float attckAmount)
     {
         PlayAnimationByValue(Constants.Animation.ENEMY_ANIMATION_TAKE_DAMAGE);
         return base.TakeDamage(attckAmount);
